ssc_pipeline/requirements.txt

pandas>=1.5
numpy>=1.23
requests>=2.31
nbformat>=5.9
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
statsmodels>=0.14
beautifulsoup4>=4.12
pyyaml>=6.0
